backend/src/soccer_analytics/api/app.py
from __future__ import annotations

from dotenv import load_dotenv
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware

from soccer_analytics.api.routes import router


def create_app() -> FastAPI:
    load_dotenv()

    app = FastAPI(title="Soccer Analytics API", version="0.1.0")

    app.add_middleware(
        CORSMiddleware,
        allow_origins=["http://localhost:5173"],
        allow_credentials=True,
        allow_methods=["*"],
        allow_headers=["*"],
    )

    app.include_router(router)

    return app


app = create_app()
