backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soccer-analytics"
version = "0.1.0"
description = "Soccer video analysis with player detection, tracking, and team classification"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.3.0",
    "supervision>=0.25.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "inference>=0.30.0",
    "pyyaml>=6.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
soccer-analytics = "soccer_analytics.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
